+++ NEWS.md
@@ -2,6 +2,10 @@
 ## unreleased
 - Use dpkg options --force-confdef and --force-confold when installing
   packages using the apt provider
+- Add support for OpenBSD
+  - Add OpenBSD package provider
+  - Add OpenBSD user provider
+  - Add OpenBSD group provider
 
 ## 0.5.0 (2015-09-21)
 - Add attribute hash support to DSL and resources